refactor: Utils에서 Splitter 제거
