+++ src/main/java/calculator/Application.java
@@ -1,7 +1,5 @@
 package calculator;
 
-import calculator.utils.Splitter;
-
 public class Application {
     public static void main(String[] args) {
         Splitter splitter = new Splitter();
